requirements.txt
bcrypt==4.0.1
click==8.0.4
dataclasses==0.8
Flask==2.0.3
Flask-Bcrypt==1.0.1
Flask-Login==0.5.0
gunicorn==21.2.0
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
packaging==21.3
psycopg2-binary==2.9.5
pyparsing==3.1.4
typing_extensions==4.1.1
Werkzeug==2.0.3
zipp==3.6.0
